+++ GL/ColoredVertex.h
@@ -10,7 +10,7 @@ namespace GL {
 class ColoredVertex : public Vertex
 {
 public:
-    ColoredVertex(const real32 &x = 0.0, const real32 &y = 0.0, const real32 &z = 0.0,
+    ColoredVertex(const real64 &x = 0.0, const real64 &y = 0.0, const real64 &z = 0.0,
            const Color &color = Color(Color::BLUE));
     ColoredVertex &setColor(const Color &new_color) {_color = new_color; return *this;}
     Color getColor() const {return _color;}
