+++ GL/TexturedVertex.cpp
@@ -2,19 +2,19 @@
 
 namespace GL {
 
-TexturedVertex::TexturedVertex(const real32 &x, const real32 &y, const real32 &z,
-                               const real32 &x_texture, const real32 &y_texture) :
+TexturedVertex::TexturedVertex(const real64 &x, const real64 &y, const real64 &z,
+                               const real64 &x_texture, const real64 &y_texture) :
     Vertex (x, y, z),
-    _x_texture(x_texture),
-    _y_texture(y_texture)
+    _x_texture(REAL32(x_texture)),
+    _y_texture(REAL32(y_texture))
 {
 
 }
 
-Vertex &TexturedVertex::setTextureCoords(const real32 &new_x, const real32 &new_y)
+Vertex &TexturedVertex::setTextureCoords(const real64 &new_x, const real64 &new_y)
 {
-    _x_texture = new_x;
-    _y_texture = new_y;
+    _x_texture = REAL32(new_x);
+    _y_texture = REAL32(new_y);
 
     return *this;
 }
